Finished serial interface
The serial interface is now working fine.
The only known issue (which is a well-known one for serial
implementations made with interrupts on AVR boards) is that if you send
too many bytes in one stroke, you will loose some of them.
After some empiric tests, the baud rate was changed from 19200 to 57600.
At 19200, a maximum of 29 bytes can be sent without losses. At 57600,
you can send up to 40 bytes (which is beyond the new designed size for a
packet) maintaining losses infeasible, but you can also go further than
that.