Overall refactor of the entire codebase
The entire GNU Make framework was redesigned to handle both host-side
and avr-side compilation and to place object files in a dedicated
directory, based on the target architecture in use.
A common test framework for host-side testing was added, and every test
was modified to use it.
Minor bugfixes and refactor operations were performed for project source
files