Finished parameterized AVR config implementation
Source and header templates have been written for AVR-side
configuration, as well as an utility to generate usable source and head
files given a CSV configuration which specify every field, with its type
and default value.
The mock NVM implementation for testing have been modified; the changes
made were not avoidable, as the way of using the NVM in the project has
changed. Rafactoring and modification of the temperature db module is
urgent.