Fixed temperature db module to work with NVM
The temperature db now support the new (and better) approach to the NVM
use, but there is a problem on how to keep track of the configuration
size in the computation of the db capacity. Likely, a new module which
takes care of initializing the entire NVM image is required.